Initial original-right-grip rework.
Okay, I'm not proud of this in the traditional sense.
I'm glad that the booleans are working out so well for me, but they are just hiding the mess that is the uncleaned photogrammetry mesh.

I'll give this a test print and see what it looks like.

Oh, and next time, SCALE BEFORE ADDING ALL THE BOOLEAN THINGS IN PLACE.
Otherwise you have to apply all of them, then scale the final mesh, then export.